requirements.txt
fastapi
uvicorn[standard]
oracledb
pydantic
pydantic-settings
python-jose[cryptography]